Rename loadedContracts.ts as this will hold generic Contract rpc methods
