+++ rpc/wallet.ts
@@ -16,7 +16,7 @@
 
 import { Wallet } from "../src/entity/Wallet";
 import { BaseContract } from "../src/contracts/BaseContract";
-import { LoadedContracts} from "./loadedContracts";
+import { LoadedContracts} from "./contracts";
 import { RPCMethod } from "./decorators";
 
 const loadedContracts = LoadedContracts.Instance;
